Animate tip changes
This is not good. I'm trying to do this in the standard react way,
but this forces me to apply layout styles that I'd rather not, and
creates an implicit value dependency between the css and javascript.

Alternatively, I could do all the animation in javascript, but then
I would have to fix the styles for the component. I don't see a way
out of this.